Revert move of `preprocess.py`
